pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-clone-engine"
version = "0.1.0"
description = "Modular copy-trading engine for tracking and mirroring trades from configured wallets across chains."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "LouisLetcher"}]
dependencies = [
  "web3>=6.10.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "SQLAlchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.2.1",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "tenacity>=8.2.3",
  "requests>=2.31.0",
  "alembic>=1.13.1",
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "solana>=0.30.2",
  "solders>=0.18.1",
  "base58>=2.1.1"
]

[tool.pytest.ini_options]
addopts = "-q --cov=trade_clone_engine --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["trade_clone_engine"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","B","UP","SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.setuptools.packages.find]
where = ["."]
include = ["trade_clone_engine*", "services*"]
exclude = ["tests*"]
